+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed the window to be active when a file is dropped
 - Added the elli and ellib functions
 - Added the fill function
+- Fixed to work even when audio is not installed
 - Updated the READMEs
 
 ## 1.6.7